Move googleTagManager setup script to its own file.
