+++ frontend/templates/troubleshooting/components/TagManager.tsx
@@ -2,38 +2,8 @@ import Head from 'next/head';
 import React from 'react';
 
 // This is public so no need to hide.
-const API_KEY = '59NVBFN';
 const GET_PARAM = 'matomoTagManager';
 const DEFAULT_ENV = 'Live';
-
-function GoogleScript() {
-   const scriptTag = `
-    (function(w,d,s,l,i) {
-      w[l]=w[l]||[];
-      w[l].push({'gtm.start': new Date().getTime(),event:'gtm.js'});
-      var f=d.getElementsByTagName(s)[0],
-      j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';
-      j.async=true;
-      j.src='https://www.googletagmanager.com/gtm.js?id='+i+dl;
-      f.parentNode.insertBefore(j,f);
-    })(window,document,'script','dataLayer','GTM-${API_KEY}');
-  `;
-
-   return <script dangerouslySetInnerHTML={{ __html: scriptTag }} />;
-}
-
-export function GoogleNoScript() {
-   return (
-      <noscript>
-         <iframe
-            src={`https://www.googletagmanager.com/ns.html?id=GTM-${API_KEY}`}
-            height="0"
-            width="0"
-            style={{ display: 'none', visibility: 'hidden' }}
-         />
-      </noscript>
-   );
-}
 
 function MatomoScript() {
    const params = new URLSearchParams();
@@ -76,7 +46,6 @@ function MatomoScript() {
 export function TagManager() {
    return (
       <Head>
-         <GoogleScript key="GoogleScript" />
          <MatomoScript key="MatomoScript" />
       </Head>
    );
